# storage/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Harency ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign in",
                "description": "Login by phone number, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone_number",
                                    "password"
                                ],
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "format": "phone_number",
                                        "example": "9786543210"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "device_token": {
                                        "type": "string",
                                        "format": "device_token",
                                        "example": "1232443bcvhsdgdfs"
                                    },
                                    "device_type": {
                                        "type": "string",
                                        "format": "device_type",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone number or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign Up",
                "description": "Sign up by phone number, password",
                "operationId": "authSignUp",
                "requestBody": {
                    "description": "Pass user details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone_number",
                                    "password",
                                    "country_code"
                                ],
                                "properties": {
                                    "phone_number": {
                                        "type": "string",
                                        "format": "phone_number",
                                        "example": "9786543210"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "device_token": {
                                        "type": "string",
                                        "format": "device_token",
                                        "example": "1232443bcvhsdgdfs"
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "format": "country_code",
                                        "example": "91"
                                    },
                                    "device_type": {
                                        "type": "string",
                                        "format": "device_type",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong phone number or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Logout user and invalidate token",
                "operationId": "authLogout",
                "requestBody": {
                    "description": "Pass user details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "device_token",
                                    "device_type"
                                ],
                                "properties": {
                                    "device_token": {
                                        "type": "string",
                                        "format": "device_token",
                                        "example": "1232443bcvhsdgdfs"
                                    },
                                    "device_type": {
                                        "type": "string",
                                        "format": "device_type",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}